Rename `src/sink.rs` to `sink_ext.rs`
