+++ src/lib.rs
@@ -95,7 +95,7 @@ pub mod serde;
 #[cfg(windows)]
 pub mod service;
 pub mod signal;
-pub(crate) mod sink;
+pub(crate) mod sink_ext;
 #[allow(unreachable_pub)]
 pub mod sinks;
 pub mod source_sender;
